+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project follows [Semantic Versioning](http://semver.org/).
 - Added an option to export LForms data in a FHIR Bundle with a DiagnosticReport resource and
   associated Observation resources.
 - Added supporting functions for FHIR Bundle.
+- Added new functions under LForms.Util for exporting/importing to/from FHIR resources. 
+- Added documentations for LForms.Util functions in form_definition.md.
 
 ## [12.7.1]
 ### Fixed